Setup to spit out measured angle based on sensor readings. Uses cutoff of 75 for max standard deviation of measured values to use (skips data sets whose stdev > 75). This is probably too high. It passes nearly all data. (Maybe that's okay, though?)
Uses floats at the moment. Gave the ol' college try to duplicating QuickStats for integers only, but I haven't tested it yet and I doubt it works (yet).

It seems to be accurate to within a degree or two compared to the input angles. This is good enough for its primary purpose, which was avoiding large angle changes by sensing the servos angle prior to sending it any signal -- before Servo::attach(int), Servo::write(int).